+++ strings/camel_case.go
@@ -1,0 +1,8 @@
+package strings
+
+// Converts string to camel case.
+//
+// Complexity: O(n)
+func CamelCase(input string) string {
+	return internalCamelCase(input, false)
+}